gogensig:remove reduant test logic
