+++ chore/gogensig/convert/_testdata/_systopkg/conf/llcppg.symb.json
@@ -1,0 +1,7 @@
+[
+  {
+    "mangle": "funcc",
+    "c++": "funcc",
+    "go": "Funcc"
+  }
+]
