+++ chore/gogensig/convert/_testdata/_systopkg/gogensig.expect
@@ -1,0 +1,5 @@
+===== temp.go =====
+package _systopkg
+
+import _ "unsafe"
+
