+++ chore/gogensig/convert/_testdata/_systopkg/hfile/temp.h
@@ -1,0 +1,16 @@
+#include <fenv.h>
+#include <limits.h>
+#include <stdarg.h>
+#include <stdbool.h>
+#include <stddef.h>
+#include <stdint.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <time.h>
+
+// posix
+#include <arpa/inet.h>
+#include <net/if.h>
+
+void funcc(uint_least64_t a, uint_least64_t b);
